+++ mapping/Snakefile
