+++ mapping/extract_vcf_snps.sh
