+++ mapping/find_intersecting_snps.py
@@ -1,849 +1,925 @@
-import argparse
-import gzip
 import sys
+import os
+import gzip
+import argparse
+import numpy as np
 
-import array
 import pysam
 
+import util
+import snptable
+
+import tables
+
+MAX_SEQS_DEFAULT = 64
+MAX_SNPS_DEFAULT = 6
 
 
-MAX_WINDOW_DEFAULT=100000
-MAX_SEQS_DEFAULT=10
-
-class SNP:
-    """SNP objects hold data for a single SNP"""
-
-    def __init__(self, snp_line):
-        """
-        Initialize SNP object.
+class DataFiles(object):
+    """Object to hold names and filehandles for all input / output 
+    datafiles"""
+    
+    def __init__(self, bam_filename, is_sorted, is_paired,
+                 output_dir=None, snp_dir=None,
+                 snp_tab_filename=None, snp_index_filename=None,
+                 haplotype_filename=None, samples=None):
+        # flag indicating whether reads are paired-end
+        self.is_paired = is_paired
         
-        Parameters:
-        -----------
+        # prefix for output files
+        self.prefix = None
 
-        snp_line : str
-            Line from SNP file.
+        # name of input BAM filename
+        self.bam_filename = bam_filename        
+        # name of sorted input bam_filename
+        # (new file is created if input file is not
+        #  already sorted)
+        self.bam_sort_filename = None
+        # pysam file handle for input BAM
+        self.input_bam = None
 
-        Attributes:
-        -----------
+        # name of output keep and to.remap BAM files
+        self.keep_filename = None
+        self.remap_filename = None
 
-        pos : int
-            Genomic position of SNP.
+        # pysam file handles for output BAM filenames
+        self.keep_bam = None
+        self.remap_bam = None
 
-        alleles : list
-            List of two alleles.
-
-        ptype : str
-            Type of polymorphism (snp or indel). If there are multiple alleles
-            and one is an indel, ptype will be "indel". If the alleles are all
-            single nucleotide variants, ptype will be "snp".
-
-        max_len : int
-            Maximum allele length. If greater than one, implies an insertion.
-
-        """
-        snp_split = snp_line.strip().split()
-        self.pos = int(snp_split[0]) - 1
-        self.alleles = [snp_split[1], snp_split[2]]
-        self.ptype = "snp"
-        self.max_len = 0
-        
-        for i in range(len(self.alleles)):
-            if self.alleles[i] == "-":
-                self.alleles[i] = ""
-                self.ptype = "indel"
-            elif len(self.alleles[i]) > self.max_len:
-                self.max_len = len(self.alleles[i])
                 
-        if self.max_len > 1:
-            self.ptype = "indel"
-        
-    def add_allele(self, new_alleles):
-        """
-        Add new alleles for a snp or indel.
-        
-        Parameters:
-        -----------
+        # name of output fastq files
+        self.fastq_single_filename = None
+        self.fastq1_filename = None
+        self.fastq2_filename = None
+        self.fastq1 = None
+        self.fastq2 = None
+        self.fastq_single = None
 
-        new_alleles : list
-            List of alleles (each allele is a string like "A").
-
-        """
-        # If a string is passed, each element of the string will be added as an
-        # allele.
-        assert type(new_alleles) is list
-        for new_allele in new_alleles:
-            if new_allele == "-":
-                self.ptype = "indel"
-                new_allele = ""
-            # Only add the new allele if it doesn't already exist.
-            if not (new_allele in self.alleles):
-                self.alleles.append(new_allele)
-                if len(new_allele) > self.max_len:
-                    self.max_len = len(new_allele)
-                    
-        if self.max_len > 1:
-            self.ptype = "indel"
-
-    def shift_indel(self):
-        """
-        Currently not used anywhere.
-        """
-        self.pos += 1
-        self.max_len -= 1
-        i = 0
-        while i < len(self.alleles):
-            if len(self.alleles) <= 1:
-                self.alleles.pop(i)
-            else:
-                self.alleles[i] = self.alleles[i][1:]
-                i += 1
-        self.alleles.append("")
-
-class BamScanner:
-    """
-    Class to keep track of all the information read in from the bamfile/snpfile.
-    """
-
-    def __init__(self, is_paired_end, max_window, filename, keep_filename,
-                 remap_name, remap_num_name, fastq_names, snp_dir,
-                 max_seqs=MAX_SEQS_DEFAULT):
-        """
-        Constructor: opens files, creates initial table.
-        
-        Attributes:
-        -----------
-
-        is_paired_end : boolean
-            Boolean indicating whether input data are paired end.
-
-        snp_dir : str
-            Path to directory that contains gzipped SNP files (one per
-            chromosome).
-
-        bamfile : pysam.Samfile
-            Input bam file that we are reading.
-
-        keep_bam : pysam.Samfile
-            Output bam file of reads that do not overlap SNPs.
-
-        remap_bam : pysam.Samfile
-            Output bam file of reads that do overlap SNPs and need to be
-            remapped.
-
-        remap_num_file : gzip.GzipFile
-            File to write XXX to.
-
-        fastqs : list
-            List of gzip.GzipFile objects for the different fastqs that will
-            contain the reads to remap.
-
-        read_table : list
-            List of lists. Sublist i contains the reads whose positions are
-            [real read position] % max_window.
-
-        cur_read : pysam.XXX
-            Current read from the bam file.
-
-        end_of_file : boolean
-            Boolean indicating whether we've reached the end of the bam file.
-
-        remap_num : int
-            A counter for the number of reads to be remapped. This starts at one
-            and is incremented when a read (pair) is written to the fastq
-            file(s). TODO: Is this supposed to start at one?
-
-        ref_match : int
-            This is incremented everytime a read sequence matches a SNP
-            genotype. Note that a particular read sequence can be looked at
-            multiple times if it has multiple SNPs, so this is somewhat hard to
-            interpret.
-
-        alt_match : int
-            This is initialized but not used anywhere.
-
-        no_match : int
-            This is incremented everytime a read sequence doesn't matche a SNP
-            genotype. Note that a particular read sequence can be looked at
-            multiple times if it has multiple SNPs, so this is somewhat hard to
-            interpret.
-
-        toss : int
-            Number of reads tossed.
-
-        nosnp : int
-            Number of reads with no SNPs. If one read in a read pair has a SNP
-            and the other doesn't, both "nosnp" and "remap" (below) will be
-            incremented by one.
-
-        remap : int
-            Number of reads to remap. If one read in a read pair has a SNP and
-            the other doesn't, both "nosnp" and "remap" (below) will be
-            incremented by one.
-
-        tot : int
-            I think this is total number of reads, although it is only
-            incremented in empty_slot_single(self) so it doesn't seem to be
-            fully implemented right now.
-
-        printstats : boolean
-            Boolean for some print statements, currently not used.
-
-        num_reads : int
-            Number of reads for a given window that we have read but not yet
-            written. This number is incremented when we read in a read and
-            decremented when we pop a read out of the read table.
-
-        window_too_small : int
-            The number of reads thrown out because their CIGAR contained a run
-            of N's longer than max_window.
-
-        cur_snp : SNP
-            The current SNP to be or being parsed. 
-
-        pos : int
-            The current genomic position we are analyzing.
-
-        chr_num : int
-            Bam file ID number of the current chromosome we are analyzing.
-
-        chr_name : str
-            Name of the current chromosome we are analyzing.
-
-        max_window : int
-            Size of the window in base pairs to process reads. All of the reads
-            and SNPs within max_window base pairs are processed at once. Any
-            junction-spanning reads (i.e. with N in the cigar) that extend
-            outside of the window are thrown out.
-
-        max_seqs : int
-            Maximum number of allelic combinations of read to consider. Reads
-            that have more than this number of allelic combinations due to 
-            a large number of overlapping SNPs are discarded.
-
-        """
-        
-        self.is_paired_end = is_paired_end
-        
-        # Read in all input files and create output files
+        # name of directory to read SNPs from
         self.snp_dir = snp_dir
-        self.bamfile = pysam.Samfile(filename,"rb")
-        self.keep_bam = pysam.Samfile(keep_filename, "wb",
-                                      template=self.bamfile)
-        self.remap_bam = pysam.Samfile(remap_name, "wb", template=self.bamfile)
-        self.remap_num_file = gzip.open(remap_num_name, "w")
-        self.fastqs = [gzip.open(fqn,"w") for fqn in fastq_names]
-        try:
-            self.cur_read = self.bamfile.next()
-        except:
-            sys.stderr.write("No lines available for input")
-            return()
-        self.end_of_file = False
 
-        self.remap_num = 1
-        self.ref_match = 0
-        self.alt_match = 0
-        self.no_match = 0
-        self.toss = 0
-        self.nosnp = 0
-        self.remap = 0
-        self.tot = 0
-        self.window_too_small = 0
+        # paths to HDF5 files to read SNP info from
+        self.snp_tab_filename = snp_tab_filename
+        self.snp_index_filename = snp_index_filename
+        self.haplotype_filename = haplotype_filename
 
-        self.printstats = True
-        
-        self.num_reads = 0
+        if self.snp_tab_filename:
+            self.snp_tab_h5 = tables.openFile(snp_tab_filename, "r")
+            self.snp_index_h5 = tables.openFile(snp_index_filename, "r")
+            self.hap_h5 = tables.openFile(haplotype_filename, "r")
+        else:
+            self.snp_tab_h5 = None
+            self.snp_index_h5 = None
+            self.hap_h5 = None
 
-        self.cur_snp = None
-        
-        self.pos = self.cur_read.pos
-        self.chr_num = self.cur_read.tid
-        self.chr_name = self.bamfile.getrname(self.cur_read.tid)
-        self.max_window = max_window
-        self.max_seqs = max_seqs
+            
+        # separate input directory and bam filename
+        tokens = self.bam_filename.split("/")
+        bam_dir = "/".join(tokens[:-1])
+        filename = tokens[-1]
+
+        if output_dir is None:
+            # if no output dir specified, use same directory as input
+            # bam file
+            output_dir = bam_dir
+        else:
+            if output_dir.endswith("/"):
+                # strip trailing '/' from output dir name
+                output_dir = output_dir[:-1]
                 
-        # Initialize the read tracking tables.
-        self.read_table = [[] for x in range(self.max_window)]
-        self.indel_dict = {}
-        self.snp_table = [0 for x in range(self.max_window)]
-        self.indel_table = [[] for x in range(self.max_window)]                
-        self.end_of_snp_file = False
-        
-        # Initialize the SNP and indel tracking tables.
-        self.switch_chr()
-        
-        # Fill all tables.
-        self.fill_table()
-     
-    def fill_table(self): 
-        """
-        Fills the table of reads starting from the current position and
-        extending for the next <max_window> base pairs. The read table is a
-        list of lists of length max_window. If the position of the current read
-        is 100, the first sublist contains all of the reads at position 100, the
-        next sublist contains all of the reads at position 101, etc.
-        """
-        if self.end_of_file:
-            return()
-            
-        # For first read we need to set self.pos and initialize the SNP table.
-        if self.num_reads == 0:
-            self.pos = self.cur_read.pos
-            self.init_snp_table()
-            #self.num_reads+=1000
-            
-        while (self.cur_read.tid == self.chr_num) and \
-          (self.cur_read.pos < (self.pos + self.max_window)):
-            self.num_reads += 1
-            self.read_table[self.cur_read.pos % self.max_window].append(self.cur_read)
-           
-            # Get a new read and check for the end of the file. 
-            try:
-                self.cur_read = self.bamfile.next()
-            except:
-                self.empty_table()
-                self.end_of_file = True
-                return()
-        
-        # Check to see if we've come across a new chromosome.
-        if self.cur_read.tid != self.chr_num:
-            self.empty_table()
-            self.chr_num = self.cur_read.tid
-            try:
-                self.chr_name = self.bamfile.getrname(self.chr_num)
-            except:
-                sys.stderr.write("Problem with tid: " + str(self.chr_num) + "\n")
-                self.skip_chr()
-            self.pos = self.cur_read.pos
-            self.switch_chr()
-            self.fill_table()
-
-
-    def switch_chr(self):
-        """Switches to looking for SNPs on next chromosome."""
-        chr_match = False
-        while not chr_match and not self.end_of_file:
-            filename = "%s/%s.snps.txt.gz" % (self.snp_dir, self.chr_name)
-            try:
-                self.snpfile = gzip.open(filename)
-                sys.stderr.write("Starting on chromosome " + self.chr_name+"\n")
-                chr_match = True
-            except:
-                sys.stderr.write("SNP file for chromosome %s not found: %s\n"
-                                 "Skipping these reads.\n" %
-                                 (self.chr_name, filename))
-                self.skip_chr()
-        
-        self.end_of_snp_file = False
-        self.get_next_snp()
-
-    def init_snp_table(self):
-        """
-        Creates an empty SNP table starting from the position of the current
-        and extending max_window base pairs. The SNP table is max_window long
-        and has a zero if there are no variants overlapping a position or
-        contains a SNP object if there is variant that overlaps a given
-        position. 
-        
-        Also creates an indel table which is a list of lists of length
-        max_window.
-
-        Also creates an indel dict which is a dict whose keys are genomic
-        positions and whose values are SNP objects whose ptype is indel.
-        """
-
-        sys.stderr.write("initializing SNP table")
-        
-        # Number of SNPs in this table. I think this is total number of
-        # different alleles across the whole table. I'm not exactly sure.
-        self.num_snps = 0
-        self.indel_dict = {}
-        self.snp_table = [0 for x in range(self.max_window)]
-        self.indel_table = [[] for x in range(self.max_window)]
-        # Get SNPs in this window but skip SNPs that are upstream of the current
-        # read.
-        while not self.end_of_snp_file and self.cur_snp.pos < self.pos:
-            self.get_next_snp()
-
-        # Add SNPs downstream of the current read and within the current window.
-        while not self.end_of_snp_file and (self.cur_snp.pos < self.pos + self.max_window):
-            if self.cur_snp.ptype == "snp":
-                self.add_snp()
-            else:
-                self.add_indel()
-            self.get_next_snp()
-        
-    def add_snp(self):
-        """
-        Add a SNP to the SNP table. If the SNP table has a zero at this
-        position, the SNP object will replace the zero. If the SNP table
-        already has a SNP object at this position, the SNP will be added as new
-        alleles. 
-        """
-        cur_pos = self.cur_snp.pos % self.max_window
-        if self.snp_table[cur_pos] == 0:
-            self.num_snps += 1
-            self.snp_table[cur_pos] = self.cur_snp
-        elif isinstance(self.snp_table[cur_pos], SNP):
-            self.snp_table[cur_pos].add_allele(self.cur_snp.alleles)     
-            
-    def add_indel(self):
-        """
-        Add an indel to the indel table and indel dict. If there is already an
-        indel in the indel dict at this position, add the alleles from cur_snp.
-        """
-        position = self.cur_snp.pos
-        if self.indel_dict.has_key(position):
-            start = self.indel_dict[position].max_len
-            self.indel_dict[position].add_allele(self.cur_snp.alleles)
+        name_split = filename.split(".")
+        if len(name_split) > 1:
+           self.prefix = output_dir + "/" + ".".join(name_split[:-1])
         else:
-            self.indel_dict[position] = self.cur_snp
-            start = 0
-        end = self.indel_dict[position].max_len
-        # max_len is the length of the longest allele for an indel and
-        # "position" is the genomic position of this indel. If the indel_dict
-        # already has an indel at this genomic position, we will append
-        # "position" to all of the positions/sublists in indel_table beyond the
-        # lenght of the indel that already exists. If there isn't already an
-        # indel in indel_table at this "position", we'll append "position" to
-        # all of the sublists in indel_table that are spanned by the indel.
-        i = start
-        while (i < end) and ((self.cur_snp.pos + i) < (self.pos + self.max_window)):
-            self.indel_table[(self.cur_snp.pos + i) % self.max_window].append(position)
-            i += 1
-
-    def get_next_snp(self):
-        """Read in next SNP (and set self.cur_snp) or signal end of file."""
-        snp_line = self.snpfile.readline()
-        if snp_line:
-            self.cur_snp = SNP(snp_line)
-        else:
-            self.end_of_snp_file = True
-
-    def skip_chr(self):
-        """Skips all of the reads from the chromosome of the current read and
-        moves on to the next chromosome. Used if the SNP file can't be
-        located."""
-        while self.cur_read.tid == self.chr_num:
-            try:
-                self.cur_read = self.bamfile.next()
-            except:
-                self.empty_table()
-                self.end_of_file = True
-                return
-
-        self.chr_num = self.cur_read.tid
-        try:
-            self.chr_name = self.bamfile.getrname(self.chr_num)
-        except:
-            sys.stderr.write("Problem with tid: " + str(self.chr_num) + "\n")
-            self.skip_chr()
-
-    def empty_slot_single(self):
-        """Processes all reads that map to the current position and
-        removes them from the read table Treats reads as single-end"""        
-        cur_slot = self.pos % self.max_window
-        while len(self.read_table[cur_slot]) > 0:
-            self.tot += 1
-            read = self.read_table[cur_slot].pop()
-            self.num_reads -= 1
-            seqs = self.check_for_snps(read, 0)
-            # num_seqs it the numbers of different sequences for this read which
-            # includes the original sequence as well as the different sequences
-            # with alternate alleles swapped in.
-            num_seqs = len(seqs)
-            if (num_seqs == 0) or (num_seqs > self.max_seqs):
-                continue
-            if (num_seqs == 1):
-                self.keep_bam.write(read)
-            else:
-                self.remap_num_file.write("%i\n" % (num_seqs - 1))
-                self.remap_num_file.flush()
-                self.remap_bam.write(read)
-                for seq in seqs[1:]:
-                    loc_line = "%i:%s:%i:%i" % (
-                        self.remap_num, 
-                        self.chr_name, 
-                        read.pos,
-                        num_seqs - 1)
-                    self.fastqs[0].write("@%s\n%s\n+%s\n%s\n" % (
-                        loc_line, 
-                        seq, 
-                        loc_line,
-                        read.qual))
-                self.remap_num += 1
-        self.shift_SNP_table()
-
-    def empty_slot_paired(self):
-        """Processes all reads that map to the current position and
-        removes them from the read table. Treats reads as paired-end."""
-        
-        cur_slot = self.pos % self.max_window
-
-        
-        # While there are reads in this slot...
-        while len(self.read_table[cur_slot]) > 0:
-            sys.stderr.write("current slot: %d\n" % cur_slot)
-
-            # Pop the first read in the slot
-            read = self.read_table[self.pos % self.max_window].pop()
-            self.num_reads -= 1
-
-            sys.stderr.write("read: %s\n" % repr(read))
+            self.prefix = output_dir + "/" + name_split[0]
             
-            # Figure out the matching read position
-            pair_chr_num = read.rnext 
-            pair_pos = read.mpos 
-            if (pair_chr_num != self.chr_num) or \
-              ((pair_pos - self.pos) > self.max_window):
-                continue
-
-            # Find the slot the matching read is in
-            pair_slot = pair_pos % self.max_window
-            
-            for indx in range(len(self.read_table[pair_slot])):
-                if self.read_table[pair_slot][indx].qname.split(":")[-1] == read.qname.split(":")[-1]:
-                    pair_read = self.read_table[pair_slot].pop(indx)
-                    self.num_reads -= 1
-                    seq1s = self.check_for_snps(read, 0)
-                    seq2s = self.check_for_snps(pair_read, read.mpos - read.pos)
-                    num_seqs = len(seq1s)*len(seq2s)
-                    if (num_seqs == 0) or (num_seqs > 32):
-                        break
-                    if (num_seqs == 1):
-                        self.keep_bam.write(read)
-                        self.keep_bam.write(pair_read)
-                    else:
-                        self.remap_bam.write(read)
-                        self.remap_bam.write(pair_read)
-                        self.remap_num_file.write("%i\n" % (2*(num_seqs-1)))
-                        first = True
-                        for seq1 in seq1s:
-                            for seq2 in seq2s:
-                                if not first:
-                                    left_pos = min(read.pos, pair_read.pos)
-                                    right_pos = max(read.pos, pair_read.pos)
-                                    loc_line="%i:%s:%i:%i:%i" % (
-                                        self.remap_num,
-                                        self.chr_name,
-                                        left_pos,
-                                        right_pos,
-                                        num_seqs - 1)
-                                    self.fastqs[0].write("@%s\n%s\n+%s\n%s\n" % (
-                                        loc_line,
-                                        seq1,
-                                        loc_line,
-                                        read.qual))
-                                    self.fastqs[1].write("@%s\n%s\n+%s\n%s\n" % (
-                                        loc_line,
-                                        self.reverse_complement(seq2),
-                                        loc_line,
-                                        pair_read.qual))
-                                first=False
-                        self.remap_num+=1
-                    # Stop searching for the pair since it was found.
-                    break 
-        self.shift_SNP_table()
-
-    def check_for_snps(self, read, start_dist):
-        """
-        Checks a single aligned read for overlapping SNPs and creates
-        alternative sequences for remapping.
-
-        Parameters
-        ----------
-        read : pysam.AlignedRead
-            Read to check for SNPs in.
-
-        start_dist : int
-            I think this is the distance from the current position of the
-            BamScanner to the start of the read.
-
-        Returns
-        -------
-        seqs : list
-            List of read sequences. This first entry is the read sequence from
-            the bam file. Any subsequent sequences are the read sequence from
-            the bam file except one base that overlapped a SNP is switched to
-            the other allele. If the list is empty, the read overlaps an indel
-            or has a CIGAR character besides N or M so we throw it out.
-        """
-        sys.stderr.write("checking read for SNP overlap\n")
+        # TODO: could allow names of output files to be specified
+        # on command line rather than appending name to prefix
+        sys.stderr.write("prefix: %s\n" % self.prefix)
         
-        indx = read.pos % self.max_window
-        # p keeps track of the number of read bases we've already analyzed. When
-        # p = length of the read, we are done with this read.
-        p = 0
-        # num_snps is the number of SNPs in this read.
-        num_snps = 0
-        # I think seg_len is the distance from the current position of the
-        # BamScanner to where we are 
-        seg_len = start_dist
-        seqs = [read.seq]
-        if start_dist > 0:
-            # has_junc indicates whether the read has an N in the CIGAR although
-            # this doesn't seem to be used anywhere.
-            has_junc = False
-        # read.cigar is a list of tuples. Each tuple has two entries. The first
-        # entry specifies the character in the cigar and the second entry
-        # specifies the length of that character. The values are
-        # M       BAM_CMATCH      0
-        # I       BAM_CINS        1
-        # D       BAM_CDEL        2
-        # N       BAM_CREF_SKIP   3
-        # S       BAM_CSOFT_CLIP  4
-        # H       BAM_CHARD_CLIP  5
-        # P       BAM_CPAD        6
-        # =       BAM_CEQUAL      7
-        # X       BAM_CDIFF       8
-        # So a tuple (0, 5) means five matches and (4, 2) means a soft clip of
-        # two.
-
-        # We'll go through each cigar tuple one at a time.
-        for cigar in read.cigar:
-            seg_len += cigar[1]
-            # Check whether this cigar segment is longer than the max window.
-            # This generally happens if there is a junction read longer than the
-            # max window.
-            if seg_len > self.max_window:
-                self.window_too_small += 1
-                return([])
-
-            if cigar[0] == 4:
-                # CIGAR indicates a soft-clipping
-                p = p + cigar[1]
-            elif cigar[0] == 0:
-                # CIGAR indicates a match alignment to the reference genome.
-                # Since there is a match, let's go through each matched base and
-                # see whether it contains a SNP.
-                for i in range(cigar[1]):  
-                    if len(self.indel_table[indx]) == 0:
-                        # sys.stderr.write("READ overlaps SNP\n")
-                        snp = self.snp_table[indx]
-                        if snp != 0:
-                            num_snps += 1
-                            if num_snps > 10:
-                                # If there are more than 10 snps overlapping,
-                                # throw out the read to prevent memory blow-up.
-                                # TODO: should we increment self.toss here?
-                                return([])
-                            for seq in list(seqs):
-                                matches = 0
-
-                                for geno in snp.alleles:
-                                    if seq[p] == geno:
-                                        matches += 1
-                                        for alt_geno in snp.alleles:
-                                            if not alt_geno == geno:
-                                                new_seq = (seq[:p] + alt_geno +
-                                                           seq[p+1:])
-                                                seqs.append(new_seq)
-                                if matches == 0:
-                                    self.no_match += 1
-                                else:
-                                    self.ref_match += 1
-                    else:
-                        # sys.stderr.write("READ overlaps INDEL\n")
-                        # It's an indel, throw it out.
-                        self.toss += 1
-                        return([])
-                    indx = (indx + 1) % self.max_window
-                    p += 1
-            elif cigar[0] == 3:
-                # Skipped in the reference genome (splice junction).
-                indx = (indx + cigar[1]) % self.max_window
-                has_junc = True
-            else:
-                # There is a non-N/M in the read CIGAR--throw out the read.
-                self.toss += 1
-                return([])
-                
-        if len(seqs) == 1:
-            self.nosnp += 1
+        if not is_sorted:
+            util.sort_bam(self.bam_filename, self.prefix)
+            self.bam_sort_filename = self.prefix + ".sort.bam"
         else:
-            self.remap += 1
-        return seqs
+            self.bam_sort_filename = self.bam_filename
+
+        self.keep_filename = self.prefix + ".keep.bam"
+        self.remap_filename = self.prefix + ".to.remap.bam"
+
+        sys.stderr.write("reading reads from:\n  %s\n" %
+                         self.bam_sort_filename)
+        
+        sys.stderr.write("writing output files to:\n")
+        
+        if self.is_paired:
+            self.fastq1_filename = self.prefix + ".remap.fq1.gz"
+            self.fastq2_filename = self.prefix + ".remap.fq2.gz"
+            self.fastq1 = gzip.open(self.fastq1_filename, "wb")
+            self.fastq2 = gzip.open(self.fastq2_filename, "wb")
+            self.fastq_single_filename = self.prefix + ".remap.single.fq.gz"
+            self.fastq_single = gzip.open(self.fastq_single_filename, "wb")
+            sys.stderr.write("  %s\n  %s\n  %s\n" %
+                             (self.fastq1_filename,
+                              self.fastq2_filename,
+                              self.fastq_single_filename))
+            
+        else:
+            self.fastq_single_filename = self.prefix + ".remap.fq.gz"
+            self.fastq_single = gzip.open(self.fastq_single_filename, "wb")
+            sys.stderr.write("  %s\n" % (self.fastq_single_filename))
+
+        self.input_bam = pysam.Samfile(self.bam_sort_filename, "rb")
+        self.keep_bam = pysam.Samfile(self.keep_filename, "wb",
+                                      template=self.input_bam)
+        self.remap_bam = pysam.Samfile(self.remap_filename, "wb",
+                                       template=self.input_bam)
+        sys.stderr.write("  %s\n  %s\n" % (self.keep_filename,
+                                           self.remap_filename))
+
+
     
-    def shift_SNP_table(self):             
-        """Shifts the SNP table over one position and makes sure that
-        indels are not lost."""
         
-        self.pos += 1
+    def close(self):
+        """close open filehandles"""
+        filehandles = [self.keep_bam, self.remap_bam, self.fastq1,
+                       self.fastq2, self.fastq_single,
+                       self.snp_tab_h5, self.snp_index_h5,
+                       self.hap_h5]
 
-        # Current slot to fill is the position + max_window - 1
-        cur_slot=(self.pos-1) % self.max_window
+        for fh in filehandles:
+            if fh:
+                fh.close()
 
-        # Delete indels that are no longer used (if they ended at the previous position)
-        for indel_pos in self.indel_table[cur_slot]:
-            if (indel_pos + self.indel_dict[indel_pos].max_len-1) == (self.pos-1):
-                del self.indel_dict[indel_pos]
         
-        self.indel_table[cur_slot]=[]
+class ReadStats(object):
+    """Track information about reads and SNPs that they overlap"""
+
+    def __init__(self):
+        # number of read matches to reference allele
+        self.ref_count = 0
+        # number of read matches to alternative allele
+        self.alt_count = 0
+        # number of reads that overlap SNP but match neither allele
+        self.other_count = 0
+
+        # number of reads discarded because not proper pair
+        self.discard_improper_pair = 0
+
+        # paired reads map to different chromosomes
+        self.discard_different_chromosome = 0
+
+        # number of reads discarded because overlap an indel
+        self.discard_indel = 0
+
+        # number of reads discarded because secondary match
+        self.discard_secondary = 0
+
+        # number of reads discarded because of too many overlapping SNPs
+        self.discard_excess_snps = 0
         
-        # Carry over indels from the previous slot
-        for indel_pos in self.indel_table[cur_slot-1]:
-            if (indel_pos + self.indel_dict[indel_pos].max_len-1) >= (self.pos+self.max_window-1):
-                self.indel_table[cur_slot].append(indel_pos)
+        # number of reads discarded because too many allelic combinations
+        self.discard_excess_reads = 0
+        
+        # reads where we expected to see other pair, but it was missing
+        # possibly due to read-pairs with different names
+        self.discard_missing_pair = 0
+        
+        # number of single reads kept
+        self.keep_single = 0
+        # number of read pairs kept
+        self.keep_pair = 0
 
-        if self.snp_table[cur_slot] != 0:
-            self.num_snps -= 1
-        self.snp_table[cur_slot] = 0
+        # number of single reads that need remapping
+        self.remap_single = 0
+        # number of read pairs kept
+        self.remap_pair = 0
+        
 
-        # See if there is a SNP overlapping the current spot.
-        while not self.end_of_snp_file and self.pos + self.max_window-1 > self.cur_snp.pos:
-            sys.stderr.write(str(self.num_snps) + " " + str(self.pos) + " " + 
-                             str(self.cur_snp.pos)+" !!!\n")
-            sys.stderr.write("SNP out of order has been skipped\n")
-            self.get_next_snp()
+    def write(self, file_handle):
+        sys.stderr.write("DISCARD reads:\n"
+                         "  improper pair: %d\n"
+                         "  different chromosome: %d\n"
+                         "  indel: %d\n"
+                         "  secondary alignment: %d\n"
+                         "  excess overlapping snps: %d\n"
+                         "  excess allelic combinations: %d\n"
+                         "  missing pairs (e.g. mismatched read names): %d\n"
+                         "KEEP reads:\n"
+                         "  single-end: %d\n"
+                         "  pairs: %d\n"
+                         "REMAP reads:\n"
+                         "  single-end: %d\n"
+                         "  pairs: %d\n" %
+                         (self.discard_improper_pair,
+                          self.discard_different_chromosome,
+                          self.discard_indel,
+                          self.discard_secondary,
+                          self.discard_excess_snps,
+                          self.discard_excess_reads,
+                          self.discard_missing_pair,
+                          self.keep_single,
+                          self.keep_pair,
+                          self.remap_single,
+                          self.remap_pair))
 
-        while not self.end_of_snp_file and (self.cur_snp.pos == (self.pos + self.max_window - 1)):
-            if self.cur_snp.ptype == "snp":
-                self.add_snp()
-            else:
-                self.add_indel()
-                if not self.cur_snp.pos in self.indel_table[cur_slot]:
-                    self.indel_table[cur_slot].append(cur_snp.pos)
-            self.get_next_snp()
-
-    def empty_table(self):
-        """Completely empties the read_table by repeatedly calling
-        empty_slot function"""
-        end_pos = self.pos + self.max_window
-        while self.pos < end_pos:
-            if self.is_paired_end:
-                self.empty_slot_paired()
-            else:
-                self.empty_slot_single()
-
-    def complement(self, letter):
-        if letter == 'A':
-            return('T')
-        elif letter == 'T':
-            return('A')
-        elif letter == 'C':
-            return('G')
-        elif letter == 'G':
-            return('C')
-        else:
-            return(letter)
-
-    def reverse_complement(self, read):
-        # reverse = ""
-        # for letter in read:
-        #     reverse = self.complement(letter) + reverse
-        # return reverse
-        reverse = [self.complement(letter) for letter in list(read)]
-        reverse.reverse()
-        return ''.join(reverse)
+        file_handle.write("read SNP ref matches: %d\n" % self.ref_count)
+        file_handle.write("read SNP alt matches: %d\n" % self.alt_count)
+        file_handle.write("read SNP mismatches: %d\n" % self.other_count)
+        
+        total = self.ref_count + self.alt_count + self.other_count
+        if total > 0:
+            mismatch_pct = 100.0 * float(self.other_count) / total
+            if mismatch_pct > 10.0:
+                sys.stderr.write("WARNING: many read SNP overlaps do not match "
+                                 "either allele (%.1f%%). SNP coordinates "
+                                 "in input file may be incorrect.\n" %
+                                 mismatch_pct)
     
-    def run(self):
-        """Iterate through bam and SNP files and write output files."""
-        self.fill_table()
-        while not self.end_of_file:
-            if self.is_paired_end:
-                self.empty_slot_paired()
-            else:
-                self.empty_slot_single()
-            self.fill_table()
-     
-        if self.window_too_small > 0:
-            sys.stderr.write(
-                'Segment distance (from read pair and junction separation) was '
-                'too large for %d reads so those reads have been thrown out. '
-                'Consider increasing the max window '
-                'size.\n' % self.window_too_small)
-
-        sys.stderr.write("Finished!\n")
-        self.keep_bam.close()
-        self.remap_bam.close()
-        self.remap_num_file.close()
-        [x.close() for x in self.fastqs]
-
 
 
 
 def parse_options():
-    parser=argparse.ArgumentParser()
+    
+    parser = argparse.ArgumentParser(description="Looks for SNPs and indels "
+                                     "overlapping reads. If a read overlaps "
+                                     "SNPs, alternative versions of the read "
+                                     "containing different alleles are created "
+                                     "and written to files for remapping. Reads "
+                                    "that do not overlap SNPs or indels are "
+                                     "written to a 'keep' BAM file."
+                                     "Reads that overlap indels are presently "
+                                     "discarded.")
+                                   
 
-    parser.add_argument("--is_paired_end", "-p", action='store_true', dest='is_paired_end', 
-                        default=False, help=("Indicates that reads are paired-end "
-                                             "(default is single)."))
+    parser.add_argument("--is_paired_end", "-p", action='store_true',
+                        dest='is_paired_end', 
+                        default=False,
+                        help=("Indicates that reads are paired-end "
+                              "(default is single)."))
     
-    parser.add_argument("--is_sorted", "-s", action='store_true', dest='is_sorted', 
-                        default=False, help=('Indicates that the input bam file'
-                                             ' is coordinate sorted (default '
-                                             'is False).'))
+    parser.add_argument("--is_sorted", "-s", action='store_true',
+                        dest='is_sorted', 
+                        default=False,
+                        help=('Indicates that the input BAM file'
+                              ' is coordinate-sorted (default '
+                              'is False).'))
     
-    mhelp = ('Changes the maximum window to search for SNPs.  The default is '
-             '%d base pairs.  Reads or read pairs that span more than this '
-             'distance (usually due to splice junctions) will be thrown out. '
-             'Increasing this window allows for longer junctions, but may '
-             'increase run time and memory requirements.' % MAX_WINDOW_DEFAULT)
-    
-    parser.add_argument("--max_window", "-m", action='store',
-                        dest='max_window', type=int, 
-                        default=MAX_WINDOW_DEFAULT, help=mhelp)
-
     parser.add_argument("--max_seqs", type=int, default=MAX_SEQS_DEFAULT,
                         help="The maximum number of sequences with different "
-                        "allelic combinations to consider remapping (default=%d). "
-                        "Reads with more allelic combinations than MAX_SEQs are discarded"
-                        % MAX_SEQS_DEFAULT)
+                        "allelic combinations to consider remapping "
+                        "(default=%d). Read pairs with more allelic combinations "
+                        "than MAX_SEQs are discarded" % MAX_SEQS_DEFAULT)
+
+    parser.add_argument("--max_snps", type=int, default=MAX_SNPS_DEFAULT,
+                        help="The maximum number of SNPs allowed to overlap "
+                        "a read before discarding the read. Allowing higher "
+                        "numbers will decrease speed and increase memory "
+                        "usage (default=%d)."
+                         % MAX_SNPS_DEFAULT)
+    
+    parser.add_argument("--output_dir", default=None,
+                        help="Directory to write output files to. If not "
+                        "specified, output files are written to the "
+                        "same directory as the input BAM file.")
+
+    parser.add_argument("--snp_dir", action='store', 
+                        help=("Directory containing SNP text files "
+                              "This directory should contain one file per "
+                              "chromosome named like chr<#>.snps.txt.gz. "
+                              "Each file should contain 3 columns: position "
+                              "RefAllele AltAllele"),
+                        default=None)
+        
+
+    parser.add_argument("--snp_tab",
+                        help="Path to HDF5 file to read SNP information "
+                        "from. Each row of SNP table contains SNP name "
+                        "(rs_id), position, allele1, allele2.",
+                        metavar="SNP_TABLE_H5_FILE",
+                        default=None)
+    
+    parser.add_argument("--snp_index",
+                        help="Path to HDF5 file containing SNP index. The "
+                        "SNP index is used to convert the genomic position "
+                        "of a SNP to its corresponding row in the haplotype "
+                        "and snp_tab HDF5 files.",
+                        metavar="SNP_INDEX_H5_FILE",
+                        default=None)
+    
+    parser.add_argument("--haplotype",
+                        help="Path to HDF5 file to read phased haplotypes "
+                        "from. When generating alternative reads "
+                        "use known haplotypes from this file rather "
+                        "than all possible allelic combinations.",
+                        metavar="HAPLOTYPE_H5_FILE",
+                        default=None)
+
+    parser.add_argument("--samples",
+                        help="Use only haplotypes and SNPs that are polymorphic in these samples. "
+                        "SAMPLES can either be a comma-delimited string "
+                        "of sample names or a path to a file with one sample name "
+                        "per line (file is assumed to be whitespace-delimited and "
+                        "first column is assumed to be sample name). "
+                        "Sample names should match those present in the "
+                        "--haplotype file. Samples are ignored if no haplotype "
+                        "file is provided.",
+                        metavar="SAMPLES")
                         
-    
     parser.add_argument("bam_filename", action='store',
-                        help="Coordinate sorted bam file containing reads.")
-
-    parser.add_argument("snp_dir", action='store', 
-                        help=('Directory containing the SNPs segregating within the '
-                              'sample in question (which need to be checked for '
-                              'mappability issues).  This directory should contain '
-                              'sorted files of SNPs separated by chromosome and named: '
-                              'chr<#>.snps.txt.gz. These files should contain 3 columns: '
-                              'position RefAllele AltAllele'))
+                        help="Coordinate-sorted input BAM file "
+                        "containing mapped reads.")
     
-    return parser.parse_args()
+        
+    options = parser.parse_args()
 
-
-
-
-def main(bam_filename, snp_dir, max_window=MAX_WINDOW_DEFAULT,
-         is_paired_end=False, is_sorted=False, max_seqs=MAX_SEQS_DEFAULT):
+    if options.snp_dir:
+        if(options.snp_tab or options.snp_index or options.haplotype):
+            parser.error("expected --snp_dir OR (--snp_tab, --snp_index and "
+                         "--haplotype) arguments but not both")
+    else:
+        if not (options.snp_tab and options.snp_index and options.haplotype):
+            parser.error("either --snp_dir OR (--snp_tab, "
+                         "--snp_index AND --haplotype) arguments must be "
+                         "provided")
     
-    name_split = bam_filename.split(".")
+    if options.samples and not options.haplotype:
+        # warn because no way to use samples if haplotype file not specified
+        sys.stderr.write("WARNING: ignoring --samples argument "
+                         "because --haplotype argument not provided")
+
+    return options
+
+
+
+
+
+
+
+
+
+def write_read(read, snp_tab, snp_idx, read_pos):
+    snp_allele1 = [' '] * read.qlen
+    snp_allele2 = [' '] * read.qlen
+
+    for (s_idx, r_idx) in zip(snp_idx, read_pos):
+        a1 = snp_tab.snp_allele1[s_idx]
+        a2 = snp_tab.snp_allele2[s_idx]
+
+        snp_allele1[r_pos-1] = a1
+        snp_allele2[r_pos-1] = a2
+
+    sys.stderr.write("READ: %s\n" % read.query)
+    sys.stderr.write("A1:   %s\n" % "".join(snp_allele1))
+    sys.stderr.write("A2:   %s\n" % "".join(snp_allele2))
     
-    if len(name_split) > 1:
-        pref = ".".join(name_split[:-1])
-    else:
-        pref = name_split[0]
-      
-    if not is_sorted:
-        pysam.sort(bam_filename, pref + ".sort")
-        sort_filename = pref + ".sort.bam"
-    else:
-        sort_filename = bam_filename
+    
+        
 
-    keep_filename = pref + ".keep.bam"
-    remap_name = pref + ".to.remap.bam"
-    remap_num_name = pref + ".to.remap.num.gz"
 
-    if is_paired_end:
-        fastq_names = [pref + ".remap.fq1.gz",
-                       pref + ".remap.fq2.gz"]
-    else:
-        fastq_names = [pref + ".remap.fq.gz"]
+def count_ref_alt_matches(read, read_stats, snp_tab, snp_idx, read_pos):
+    ref_alleles = snp_tab.snp_allele1[snp_idx]
+    alt_alleles = snp_tab.snp_allele2[snp_idx]
+    
+    for i in range(len(snp_idx)):
+        if ref_alleles[i] == read.query[read_pos[i]-1]:
+            # read matches reference allele
+            read_stats.ref_count += 1
+        elif alt_alleles[i] == read.query[read_pos[i]-1]:
+            # read matches non-reference allele
+            read_stats.alt_count += 1
+        else:
+            # read matches neither ref nor other
+            read_stats.other_count += 1
+            
 
-    bam_data = BamScanner(is_paired_end, max_window, 
-                          sort_filename, keep_filename, remap_name, 
-                          remap_num_name, fastq_names, snp_dir,
-                          max_seqs=max_seqs)
-    bam_data.run()
+
+def get_unique_haplotypes(haplotypes, snp_idx):
+    """returns list of vectors of unique haplotypes for this set of SNPs"""
+    haps = haplotypes[snp_idx,:].T
+    
+    # create view of data that joins all elements of column
+    # into single void datatype
+    h = np.ascontiguousarray(haps).view(np.dtype((np.void, haps.dtype.itemsize * haps.shape[1])))
+
+    # get index of unique columns
+    _, idx = np.unique(h, return_index=True)
+
+    return haps[idx,:]
+    
+
+
+
+            
+def generate_haplo_reads(read_seq, snp_idx, read_pos, ref_alleles, alt_alleles,
+                         haplo_tab):
+    haps = get_unique_haplotypes(haplo_tab, snp_idx)
+
+    # sys.stderr.write("UNIQUE haplotypes: %s\n"
+    #                  "read_pos: %s\n"
+    #                 % (repr(haps), read_pos))
+    
+    read_len = len(read_seq)
+
+    new_read_list = []
+
+    # loop over haplotypes
+    for hap in haps:
+        new_read = []
+        cur_pos = 1
+
+        missing_data = False
+
+        # loop over the SNPs to get alleles that make up this haplotype
+        for i in range(len(hap)):
+            if read_pos[i] > cur_pos:
+                # add segment of read
+                new_read.append(read_seq[cur_pos-1:read_pos[i]-1])
+            # add segment for appropriate allele
+            if hap[i] == 0:
+                # reference allele
+                new_read.append(ref_alleles[i])
+            elif hap[i] == 1:
+                # alternate allele
+                new_read.append(alt_alleles[i])
+            else:
+                # haplotype has unknown genotype or phasing so skip it...
+                # not sure if this is the best thing to do, could instead
+                # assume ambiguity of this allele and generate reads with
+                # both possible alleles
+                missing_data = True
+                break
+            
+            cur_pos = read_pos[i] + 1
+
+        if read_len >= cur_pos:
+            # add final segment
+            new_read.append(read_seq[cur_pos-1:read_len])
+            
+        if not missing_data:
+            new_seq = "".join(new_read)
+
+            # sanity check: read should be same length as original
+            if len(new_seq) != read_len:
+                raise ValueError("Expected read len to be %d but "
+                                 "got %d.\n"
+                                 "ref_alleles: %s\n"
+                                 "alt_alleles: %s\n"
+                                 "read_pos: %s\n"
+                                 "snp_idx: %s\n"
+                                 "haps: %s\n" 
+                                 % (read_len, len(new_seq),
+                                    repr(ref_alleles), repr(alt_alleles),
+                                    repr(read_pos), repr(snp_idx),
+                                    repr(haps)))
+
+            new_read_list.append("".join(new_seq))
+
+    return new_read_list
+
+    
+
+            
+def generate_reads(read_seq, read_pos, ref_alleles, alt_alleles, i):
+    """Recursively generate set of reads with all possible combinations
+    of alleles (i.e. 2^n combinations where n is the number of snps overlapping
+    the reads)
+    """
+    # TODO: this would use a lot less memory if re-implemented
+    # to not use recursion
+    
+    # create new version of this read with both reference and
+    # alternative versions of allele at this index
+    idx = read_pos[i]-1
+    ref_read = read_seq[:idx] + ref_alleles[i] + read_seq[idx+1:]
+    alt_read = read_seq[:idx] + alt_alleles[i] + read_seq[idx+1:]
+
+    if i == len(read_pos)-1:
+        # this was the last SNP
+        return [ref_read, alt_read]
+
+    # continue recursively with other SNPs overlapping this read
+    reads1 = generate_reads(ref_read, read_pos, ref_alleles, alt_alleles, i+1)
+    reads2 = generate_reads(alt_read, read_pos, ref_alleles, alt_alleles,  i+1)
+
+    return reads1 + reads2
+                
+
+
+def write_fastq(fastq_file, orig_read, new_seqs):
+    n_seq = len(new_seqs)
+    i = 1
+    for new_seq in new_seqs:
+        # Give each read a new name giving:
+        # 1 - the original name of the read
+        # 2 - the coordinate that it should map to
+        # 3 - the number of the read
+        # 4 - the total number of reads being remapped
+        name = "%s.%d.%d.%d" % (orig_read.qname, orig_read.pos+1, i, n_seq)
+                                       
+        fastq_file.write("@%s\n%s\n+%s\n%s\n" %
+                         (name, new_seq, name, orig_read.qual))
+
+        i += 1
 
         
+def write_pair_fastq(fastq_file1, fastq_file2, orig_read1, orig_read2,
+                     new_pairs):
+
+    n_pair = len(new_pairs)
+    i = 1
+    for pair in new_pairs:
+        # give each fastq record a new name giving:
+        # 1 - the original name of the read
+        # 2 - the coordinates the two ends of the pair should map to
+        # 3 - the number of the read
+        # 4 - the total number of reads being remapped
+
+        pos_str = "%d-%d" % (min(orig_read1.pos+1, orig_read2.pos+1),
+                             max(orig_read1.pos+1, orig_read2.pos+1))
+        
+        name = "%s.%s.%d.%d" % (orig_read1.qname, pos_str, i, n_pair)
+        
+        fastq_file1.write("@%s\n%s\n+%s\n%s\n" %
+                          (name, pair[0], name, orig_read1.qual))
+
+        rev_seq = util.revcomp(pair[1])
+        fastq_file2.write("@%s\n%s\n+%s\n%s\n" %
+                          (name, rev_seq, name, orig_read2.qual))
+
+        i += 1
+                         
+
+
+        
+    
+def filter_reads(files, max_seqs=MAX_SEQS_DEFAULT, max_snps=MAX_SNPS_DEFAULT,
+                 samples=None):
+    cur_chrom = None
+    cur_tid = None
+    seen_chrom = set([])
+
+    snp_tab = snptable.SNPTable()
+    read_stats = ReadStats()
+    read_pair_cache = {}
+    cache_size = 0
+    read_count = 0
+    
+    for read in files.input_bam:
+        read_count += 1
+        # if (read_count % 100000) == 0:
+        #     sys.stderr.write("\nread_count: %d\n" % read_count)
+        #     sys.stderr.write("cache_size: %d\n" % cache_size)
+        
+        if (cur_tid is None) or (read.tid != cur_tid):
+            # this is a new chromosome
+            cur_chrom = files.input_bam.getrname(read.tid)
+
+            if len(read_pair_cache) != 0:
+                sys.stderr.write("WARNING: failed to find pairs for %d "
+                                 "reads on this chromosome\n" %
+                                 len(read_pair_cache))
+                read_stats.discard_missing_pair += len(read_pair_cache)
+            read_pair_cache = {}
+            cache_size = 0
+            read_count = 0
+            
+            if cur_chrom in seen_chrom:
+                # sanity check that input bam file is sorted
+                raise ValueError("expected input BAM file to be sorted "
+                                 "but chromosome %s is repeated\n" % cur_chrom)
+            seen_chrom.add(cur_chrom)
+            cur_tid = read.tid
+            sys.stderr.write("starting chromosome %s\n" % cur_chrom)
+
+            # use HDF5 files if they are provided, otherwise use text
+            # files from SNP dir
+            if files.snp_tab_h5:
+                sys.stderr.write("reading SNPs from file '%s'\n" %
+                                 files.snp_tab_h5.filename)
+                snp_tab.read_h5(files.snp_tab_h5, files.snp_index_h5,
+                                files.hap_h5, cur_chrom, samples)
+            else:
+                snp_filename = "%s/%s.snps.txt.gz" % (files.snp_dir, cur_chrom)
+                sys.stderr.write("reading SNPs from file '%s'\n" % snp_filename)
+                snp_tab.read_file(snp_filename)
+            
+            sys.stderr.write("processing reads\n")
+
+        if read.is_secondary:
+            # this is a secondary alignment (i.e. read was aligned more than
+            # once and this has align score that <= best score)
+            read_stats.discard_secondary += 1
+            continue
+
+        if read.is_paired:
+            if read.next_reference_name is None:
+                # other side of pair not mapped...
+                process_single_read(read, read_stats, files, snp_tab, max_seqs,
+                                    max_snps)
+            elif(read.next_reference_name == cur_chrom or
+                 read.next_reference_name == "="):
+                # other pair mapped to same chrom
+
+                # sys.stderr.write("flag: %s" % read.flag)
+                if not read.is_proper_pair:
+                    # sys.stderr.write(' => improper\n')
+                    read_stats.discard_improper_pair += 1
+                    continue
+                # sys.stderr.write(' => proper\n')
+
+                if read.qname in read_pair_cache:
+                    # we already saw prev pair, retrieve from cache
+                    read1 = read_pair_cache[read.qname]
+                    read2 = read
+                    del read_pair_cache[read.qname]
+                    cache_size -= 1
+
+                    if read2.next_reference_start != read1.reference_start:
+                        sys.stderr.write("WARNING: read pair positions "
+                                         "do not match for pair %s\n" %
+                                         read.qname)
+                    else:
+                        process_paired_read(read1, read2, read_stats,
+                                            files, snp_tab, max_seqs,
+                                            max_snps)
+                else:
+                    # we need to wait for next pair
+                    read_pair_cache[read.qname] = read
+
+                    cache_size += 1
+
+                    
+            else:
+                # other side of pair mapped to different
+                # chromosome, discard this read
+                read_stats.discard_different_chromosome += 1
+
+        else:
+            process_single_read(read, read_stats, files, snp_tab,
+                                max_seqs, max_snps)
+
+    if len(read_pair_cache) != 0:
+        sys.stderr.write("WARNING: failed to find pairs for %d "
+                         "reads on this chromosome\n" %
+                         len(read_pair_cache))
+        read_stats.discard_missing_pair += len(read_pair_cache)
+    
+    read_stats.write(sys.stderr)
+                     
+
+def process_paired_read(read1, read2, read_stats, files,
+                        snp_tab, max_seqs, max_snps):
+    """Checks if either end of read pair overlaps SNPs or indels
+    and writes read pair (or generated read pairs) to appropriate 
+    output files"""
+
+    new_reads = []    
+    for read in (read1, read2):
+        # check if either read overlaps SNPs or indels
+        # check if read overlaps SNPs or indels
+        snp_idx, snp_read_pos, \
+            indel_idx, indel_read_pos = snp_tab.get_overlapping_snps(read)
+        
+        if len(indel_idx) > 0:
+            # for now discard this read pair, we want to improve this to handle
+            # the indel reads appropriately
+            read_stats.discard_indel += 2
+            # TODO: add option to handle indels instead of throwing out reads
+            return
+
+        if len(snp_idx) > 0:
+            ref_alleles = snp_tab.snp_allele1[snp_idx]
+            alt_alleles = snp_tab.snp_allele2[snp_idx]
+
+            count_ref_alt_matches(read, read_stats, snp_tab, snp_idx,
+                                  snp_read_pos)
+
+            # limit recursion here by discarding reads that
+            # overlap too many SNPs
+            if len(snp_read_pos) > max_snps:
+                read_stats.discard_excess_snps += 1
+                return
+
+            if files.hap_h5:
+                # generate reads using observed set of haplotypes
+                read_seqs = generate_haplo_reads(read.query, snp_idx,
+                                                 snp_read_pos,
+                                                 ref_alleles, alt_alleles,
+                                                 snp_tab.haplotypes)
+            else:
+                # generate all possible allelic combinations of reads
+                read_seqs = generate_reads(read.query, snp_read_pos,
+                                           ref_alleles, alt_alleles, 0)
+            
+            new_reads.append(read_seqs)
+        else:
+            # no SNPs or indels overlap this read
+            new_reads.append([])
+            
+    if len(new_reads[0]) == 0 and len(new_reads[1]) == 0:
+        # neither read overlapped SNPs or indels
+        files.keep_bam.write(read1)
+        files.keep_bam.write(read2)
+        read_stats.keep_pair += 1
+    else:
+        # add original version of both sides of pair
+        new_reads[0].append(read1.query)
+        new_reads[1].append(read2.query)
+
+        if len(new_reads[0]) + len(new_reads[1]) > max_seqs:
+            # quit now before generating a lot of read pairs
+            read_stats.discard_excess_reads += 2
+            return 
+
+        # collect all unique combinations of read pairs
+        unique_pairs = set([])
+        n_unique_pairs = 0
+        for new_read1 in new_reads[0]:
+            for new_read2 in new_reads[1]:
+                pair = (new_read1, new_read2)
+                if pair in unique_pairs:
+                    pass
+                else:
+                    n_unique_pairs += 1
+                    if n_unique_pairs > max_seqs:
+                        read_stats.discard_excess_reads += 2
+                        return
+                    unique_pairs.add(pair)
+
+        # remove original read pair, if present
+        orig_pair = (read1.query, read2.query)
+                                 
+        if orig_pair in unique_pairs:
+            unique_pairs.remove(orig_pair)
+            
+        # write read pair to fastqs for remapping
+        write_pair_fastq(files.fastq1, files.fastq2, read1, read2,
+                         unique_pairs)
+
+        # Write read to 'remap' BAM for consistency with previous
+        # implementation of script. Probably not needed and will result in
+        # BAM that is not coordinate sorted. Possibly remove this...
+        files.remap_bam.write(read1)
+        files.remap_bam.write(read2)
+        read_stats.remap_pair += 1
+        
+
+        
+    
+
+    
+
+def process_single_read(read, read_stats, files, snp_tab, max_seqs,
+                        max_snps):
+    """Check if a single read overlaps SNPs or indels, and writes
+    this read (or generated read pairs) to appropriate output files"""
+                
+    # check if read overlaps SNPs or indels
+    snp_idx, snp_read_pos, \
+        indel_idx, indel_read_pos = snp_tab.get_overlapping_snps(read)
+
+    
+    if len(indel_idx) > 0:
+        # for now discard this read, we want to improve this to handle
+        # the indel reads appropriately
+        read_stats.discard_indel += 1
+        # TODO: add option to handle indels instead of throwing out reads
+        return
+
+    if len(snp_idx) > 0:
+        ref_alleles = snp_tab.snp_allele1[snp_idx]
+        alt_alleles = snp_tab.snp_allele2[snp_idx]
+
+        count_ref_alt_matches(read, read_stats, snp_tab, snp_idx,
+                              snp_read_pos)
+
+        # limit recursion here by discarding reads that
+        # overlap too many SNPs
+        if len(snp_read_pos) > max_snps:
+            read_stats.discard_excess_snps += 1
+            return
+
+        if files.hap_h5:
+            read_seqs = generate_haplo_reads(read.query, snp_idx,
+                                             snp_read_pos,
+                                             ref_alleles, alt_alleles,
+                                             snp_tab.haplotypes)
+        else:
+            read_seqs = generate_reads(read.query,  snp_read_pos,
+                                       ref_alleles, alt_alleles, 0)
+
+        # make set of unique reads, we don't want to remap
+        # duplicates, or the read that matches original
+        unique_reads = set(read_seqs)
+        if read.query in unique_reads:
+            unique_reads.remove(read.query)
+        
+        if len(unique_reads) == 0:
+            # only read generated matches original read,
+            # so keep original
+            files.keep_bam.write(read)
+            read_stats.keep_single += 1
+        elif len(unique_reads) < max_seqs:
+            # write read to fastq file for remapping
+            write_fastq(files.fastq_single, read, unique_reads)
+
+            # write read to 'to remap' BAM
+            # this is probably not necessary with new implmentation
+            # but kept for consistency with previous version of script
+            files.remap_bam.write(read)
+            read_stats.remap_single += 1
+        else:
+            # discard read
+            read_stats.discard_excess_reads += 1
+            return
+
+    else:
+        # no SNPs overlap read, write to keep file
+        files.keep_bam.write(read)
+        read_stats.keep_single += 1
+            
+
+
+
+def parse_samples(samples_str):
+    """Gets list of samples from --samples argument. This may be 
+    a comma-delimited string or a path to a file. If a file is provided 
+    then the first column of the file is assumed to be the sample name"""
+
+    if samples_str is None:
+        return None
+        
+    # first check if this is a path to a file
+    if os.path.exists(samples_str) and not os.path.isdir(samples_str):
+        samples = []
+
+        if samples_str.endswith(".gz"):
+            f = gzip.open(samples_str)
+        else:
+            f = open(samples_str)
+
+        for line in f:
+            # assume first token in line is sample name
+            samples.append(line.split()[0])
+
+        sys.stderr.write("read %d sample names from file '%s'\n" %
+                         (len(samples), samples_str))
+                    
+        f.close()
+    else:    
+        # otherwise assume comma-delimited string
+        if ("," not in samples_str and len(samples_str) > 15) \
+           or ("/" in samples_str):
+            sys.stderr.write("WARNING: --samples argument (%s) "
+                             "does not look like sample name "
+                             "but is not path to valid file. "
+                             "Assuming it is a sample name anyway."
+                             % samples_str)
+
+        samples = samples_str.split(",")
+        sys.stderr.write("SAMPLES: %s\n"% repr(samples))
+
+
+    return samples
+
+
+        
+def main(bam_filenames, is_paired_end=False,
+         is_sorted=False, max_seqs=MAX_SEQS_DEFAULT,
+         max_snps=MAX_SNPS_DEFAULT, output_dir=None,
+         snp_dir=None, snp_tab_filename=None,
+         snp_index_filename=None,
+         haplotype_filename=None, samples=None):
+
+    files = DataFiles(bam_filenames,  is_sorted, is_paired_end,
+                      output_dir=output_dir,
+                      snp_dir=snp_dir,
+                      snp_tab_filename=snp_tab_filename,
+                      snp_index_filename=snp_index_filename,
+                      haplotype_filename=haplotype_filename)
+    
+    filter_reads(files, max_seqs=max_seqs, max_snps=max_snps,
+                 samples=samples)
+
+    files.close()
+    
+    
 
 if __name__ == '__main__':
     options = parse_options()
-    
-    main(options.bam_filename, options.snp_dir, max_window=options.max_window,
-         is_paired_end=options.is_paired_end, is_sorted=options.is_sorted,
-         max_seqs=options.max_seqs)
 
+    samples = parse_samples(options.samples)
+
+    sys.stderr.write("command line: %s\n" % " ".join(sys.argv))
+    
+    main(options.bam_filename,
+         is_paired_end=options.is_paired_end, is_sorted=options.is_sorted,
+         max_seqs=options.max_seqs, max_snps=options.max_snps,
+         output_dir=options.output_dir,
+         snp_dir=options.snp_dir,
+         snp_tab_filename=options.snp_tab,
+         snp_index_filename=options.snp_index,
+         haplotype_filename=options.haplotype,
+         samples=samples)
+         
+
