+++ mapping/run_snakemake.sh
