+++ mapping/snake_conf.yaml
