+++ mapping/snptable.py
