+++ mapping/test_snptable.py
