+++ mapping/util.py
